+++ apps/bm_pwl.cc
@@ -172,10 +172,10 @@ TIME_PAIR EVAL_BM_PWL::tr_review(COMPONENT* d)const
 /*--------------------------------------------------------------------------*/
 bool EVAL_BM_PWL::parse_numlist(CS& cmd)
 {
-  unsigned start = cmd.cursor();
-  unsigned here = cmd.cursor();
+  size_t start = cmd.cursor();
+  size_t here = cmd.cursor();
   for (;;) {
-    unsigned start_of_pair = here;
+    size_t start_of_pair = here;
     std::pair<PARAMETER<double>, PARAMETER<double> > p;
     //cmd >> key >> value;
     cmd >> p.first; // key
